Preparations for creating an AddNewResource component
